+++ src/surrealist/ql/statements/alter_statements.py
@@ -1,0 +1,5 @@
+from surrealist.ql.statements.statement import FinishedStatement, Statement
+
+
+
+