+++ src/surrealist/ql/statements/define_index_statements.py
@@ -1,32 +1,10 @@
-import json
 from typing import Optional
 
+from surrealist.ql.statements.common_statements import CanUseComment
 from surrealist.ql.statements.statement import FinishedStatement, Statement
 
 
 # This module contains classes with duplicate code. This is made specifically for IDE hints
-
-class Comment(FinishedStatement):
-    """
-    Represents a COMMENT clause in DEFINE INDEX statement.
-    """
-    def __init__(self, statement: Statement, text: str):
-        super().__init__(statement)
-        self._text = text
-
-    def _clean_str(self):
-        text = f" COMMENT {json.dumps(self._text)}" if self._text else ""
-        return f"{self._statement._clean_str()}{text}"
-
-
-class CanUseComment:
-    def comment(self, comment: str) -> Comment:
-        """
-        Adds COMMENT statement to the query
-        :param comment: comment text
-        :return: self
-        """
-        return Comment(self, comment)
 
 
 class Unique(FinishedStatement, CanUseComment):