+++ ReleaseNotes.md
@@ -18,6 +18,7 @@
 - add UPSERT statement
 - add DEFINE ACCESS JWT and DEFINE ACCESS RECORD statements
 - add REMOVE ACCESS statement
+- add ALTER TABLE statement
 - deprecate DEFINE TOKEN-REMOVE TOKEN and DEFINE SCOPE-REMOVE SCOPE
 - let and unset now raises CompatibilityError for http
 - add upsert method to Connection